requirement.txt
streamlit
pandas
numpy
websocket-client
statsmodels
scipy
plotly

